+++ prefect_gcp/cloud_storage.py
@@ -472,6 +472,11 @@ class GcsBucket(WritableDeploymentStorage, WritableFileSystem, ObjectStorageBloc
     """
     Block used to store data using GCP Cloud Storage Buckets.
 
+    Note! `GcsBucket` in `prefect-gcp` is a unique block, separate from `GCS`
+    in core Prefect. `GcsBucket` does not use `gcsfs` under the hood,
+    instead using the `google-cloud-storage` package, and offers more configuration
+    and functionality.
+
     Attributes:
         bucket: Name of the bucket.
         gcp_credentials: The credentials to authenticate with GCP.
@@ -539,7 +544,7 @@ class GcsBucket(WritableDeploymentStorage, WritableFileSystem, ObjectStorageBloc
     @sync_compatible
     async def get_directory(
         self, from_path: Optional[str] = None, local_path: Optional[str] = None
-    ) -> None:
+    ) -> List[Union[str, Path]]:
         """
         Copies a folder from the configured GCS bucket to a local directory.
         Defaults to copying the entire contents of the block's bucket_folder
@@ -550,6 +555,9 @@ class GcsBucket(WritableDeploymentStorage, WritableFileSystem, ObjectStorageBloc
                 configured bucket_folder.
             local_path: Local path to download GCS bucket contents to.
                 Defaults to the current working directory.
+
+        Returns:
+            A list of downloaded file paths.
         """
         from_path = (
             self.bucket_folder if from_path is None else self._resolve_path(from_path)
@@ -558,7 +566,7 @@ class GcsBucket(WritableDeploymentStorage, WritableFileSystem, ObjectStorageBloc
         if local_path is None:
             local_path = os.path.abspath(".")
         else:
-            local_path = os.path.expanduser(local_path)
+            local_path = os.path.abspath(os.path.expanduser(local_path))
 
         project = self.gcp_credentials.project
         client = self.gcp_credentials.get_cloud_storage_client(project=project)
@@ -566,6 +574,8 @@ class GcsBucket(WritableDeploymentStorage, WritableFileSystem, ObjectStorageBloc
         blobs = await run_sync_in_worker_thread(
             client.list_blobs, self.bucket, prefix=from_path
         )
+
+        file_paths = []
         for blob in blobs:
             blob_path = blob.name
             if blob_path[-1] == "/":
@@ -575,12 +585,14 @@ class GcsBucket(WritableDeploymentStorage, WritableFileSystem, ObjectStorageBloc
             os.makedirs(os.path.dirname(local_file_path), exist_ok=True)
 
             with disable_run_logger():
-                await cloud_storage_download_blob_to_file.fn(
+                file_path = await cloud_storage_download_blob_to_file.fn(
                     bucket=self.bucket,
                     blob=blob_path,
                     path=local_file_path,
                     gcp_credentials=self.gcp_credentials,
                 )
+                file_paths.append(file_path)
+        return file_paths
 
     @sync_compatible
     async def put_directory(
@@ -754,6 +766,46 @@ class GcsBucket(WritableDeploymentStorage, WritableFileSystem, ObjectStorageBloc
 
         # Ignore folders
         return [blob for blob in blobs if not blob.name.endswith("/")]
+
+    @sync_compatible
+    async def list_folders(self, folder: str = "") -> List[str]:
+        """
+        Lists all folders and subfolders in the bucket.
+
+        Args:
+            folder: List all folders and subfolders inside given folder.
+
+        Returns:
+            A list of folders.
+
+        Examples:
+            Get all folders from a bucket named "my-bucket".
+            ```python
+            from prefect_gcp.cloud_storage import GcsBucket
+
+            gcs_bucket = GcsBucket.load("my-bucket")
+            gcs_bucket.list_folders()
+            ```
+
+            Get all folders from a folder called years
+            ```python
+            from prefect_gcp.cloud_storage import GcsBucket
+
+            gcs_bucket = GcsBucket.load("my-bucket")
+            gcs_bucket.list_folders('years)
+            ```
+        """
+
+        bucket_path = self._join_bucket_folder()
+        self.logger.info(f"Listing folders in bucket {bucket_path}.")
+
+        blobs = await self.list_blobs(folder)
+        # gets all folders with full path
+        folders = {
+            str(PurePosixPath(blob.name).parent).replace(".", "") for blob in blobs
+        }
+
+        return list(folders)
 
     @sync_compatible
     async def download_object_to_path(