+++ tests/test_aiplatform.py
@@ -1,4 +1,4 @@
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, patch
 
 import pytest
 from google.cloud.aiplatform_v1.types.job_state import JobState
@@ -21,8 +21,13 @@ class TestVertexAICustomTrainingJob:
         )
 
     # TODO: Improve test resiliency to changes in str output
-    def test_preview(self, vertex_ai_custom_training_job: VertexAICustomTrainingJob):
+    @patch("prefect_gcp.aiplatform.VertexAICustomTrainingJob._base_environment")
+    def test_preview(
+        self, mock_base_env, vertex_ai_custom_training_job: VertexAICustomTrainingJob
+    ):
+        mock_base_env.return_value = {"PREFECT_API_KEY": "secret"}
         actual_lines = vertex_ai_custom_training_job.preview().splitlines()
+
         expected_lines = """
             display_name: "container
             job_spec {
@@ -31,6 +36,10 @@ class TestVertexAICustomTrainingJob:
                         image_uri: "us-docker.pkg.dev/cloudrun/container/job:latest"
                         command: "echo"
                         command: "hello!!"
+                        env {
+                          name: "PREFECT_API_KEY"
+                          value: "secret"
+                      }
                     }
                     machine_spec {
                         machine_type: "n1-standard-4"
@@ -42,10 +51,34 @@ class TestVertexAICustomTrainingJob:
                 service_account: "my_service_account_email"
             }
         """.strip().splitlines()
+
         for actual_line, expected_line in zip(actual_lines, expected_lines):
             if '"container' in actual_line:
                 actual_line = actual_line.split("-")[0]  # remove the unique hex
             assert actual_line.strip() == expected_line.strip()  # disregard whitespace
+
+    @patch("prefect_gcp.aiplatform.VertexAICustomTrainingJob._base_environment")
+    def test_environment_variables(self, mock_base_env, gcp_credentials):
+        mock_base_env.return_value = {"PREFECT_API_KEY": "secret", "FOO": "INITIAL"}
+        vertex_job = VertexAICustomTrainingJob(
+            command=["echo", "hello!!"],
+            region="us-east1",
+            image="us-docker.pkg.dev/cloudrun/container/job:latest",
+            gcp_credentials=gcp_credentials,
+            env={"FOO": "BAR"},  # overrides
+        )
+        job_spec = vertex_job._build_job_spec()
+
+        assert len(job_spec.worker_pool_specs) == 1
+
+        env_list_in_container_spec = job_spec.worker_pool_specs[0].container_spec.env
+
+        expected_env = VertexAICustomTrainingJob._base_environment().copy()
+        expected_env.update({"FOO": "BAR", "PREFECT_API_KEY": "secret"})
+
+        for item in env_list_in_container_spec:
+            assert item.name in expected_env
+            assert item.value == expected_env[item.name]
 
     def test_kill(
         self, vertex_ai_custom_training_job: VertexAICustomTrainingJob, caplog