+++ CHANGELOG.md
@@ -17,9 +17,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Listing blobs at the root folder - [#120](https://github.com/PrefectHQ/prefect-gcp/pull/120)
-
 ### Security
+
+## 0.2.6
+
+Released on February 7th, 2023.
+
+### Fixed
+
+- `get_directory` method in `GcsBucket` returns the list of downloaded file paths and supports relative path for `local_path` - [#129](https://github.com/PrefectHQ/prefect-gcp/pull/129).
+- Reporting the state of the `VertexAICustomTrainingJob` to the Prefect API by passing the base env in job spec - [#132](https://github.com/PrefectHQ/prefect-gcp/pull/132).
+
+## 0.2.5
+
+Released on January 27th, 2023.
+
+### Added
+
+- `vpc_connector_name` field to `CloudRunJob` - [#123](https://github.com/PrefectHQ/prefect-gcp/pull/123)
+- `list_folders` method for GcsBucket - [#121](https://github.com/PrefectHQ/prefect-gcp/pull/121)
+
+### Fixed
+
+- Listing blobs at the root folder - [#120](https://github.com/PrefectHQ/prefect-gcp/pull/120)
 
 ## 0.2.4